dealer phone call added on list item call dealer click

+++ app/src/main/java/com/example/gmassignment/carListing/data/response/Listings.kt
@@ -33,7 +33,7 @@ data class Listings(
     val isEnriched: Boolean,
     val listPrice: Double,
     val make: String,
-    val mileage: Int,
+    val mileage: Long,
     val model: String,
     val monthlyPaymentEstimate: MonthlyPaymentEstimate,
     val mpgCity: Int,
